refactor(sdk): Split types.ts file into multiple organized type definition files
