+++ packages/sdk/src/types/TExistentialDeposit.ts
@@ -1,0 +1,3 @@
+import { type TNodeWithRelayChains } from './TNode'
+
+export type TEdJsonMap = Record<TNodeWithRelayChains, string | null>
