+++ packages/sdk/src/types/TRelayChain.ts
@@ -1,0 +1,2 @@
+export type TRelayChainType = 'polkadot' | 'kusama'
+export type TRelayChainSymbol = 'DOT' | 'KSM'
