+++ packages/sdk/src/types/TTransfer.ts
@@ -1,52 +1,18 @@
-//  derrived from https://github.com/kodadot/packages/blob/main/minimark/src/common/types.ts
-
 import { type ApiPromise } from '@polkadot/api'
+import { type TMultiLocation } from './TMultiLocation'
+import { type TNode } from './TNode'
 import { type SubmittableExtrinsic } from '@polkadot/api/types'
-import {
-  type NODES_WITH_RELAY_CHAINS,
-  type NODE_NAMES,
-  type SUPPORTED_PALLETS
-} from './maps/consts'
-import { type TMultiLocation } from './types/TMultiLocation'
 
-export type UpdateFunction = (name: string, index: number) => string
 export type Extrinsic = SubmittableExtrinsic<'promise'>
-export type ExtrinsicFunction<T> = (arg: T) => Extrinsic
-export type TRelayChainType = 'polkadot' | 'kusama'
-export type TRelayChainSymbol = 'DOT' | 'KSM'
-export type TNode = (typeof NODE_NAMES)[number]
-export type TNodeWithRelayChains = (typeof NODES_WITH_RELAY_CHAINS)[number]
-export interface TAssetDetails {
-  assetId: string
-  symbol?: string
-  decimals?: number
-}
-export interface TNativeAssetDetails {
-  assetId?: string
-  symbol: string
-  decimals: number
-}
-export interface TNodeAssets {
-  paraId: number
-  relayChainAssetSymbol: TRelayChainSymbol
-  nativeAssetSymbol: string
-  nativeAssets: TNativeAssetDetails[]
-  otherAssets: TAssetDetails[]
-}
-export type TAssetJsonMap = Record<TNode, TNodeAssets>
-export type TScenario = 'ParaToRelay' | 'ParaToPara' | 'RelayToPara'
-export type TPallet = (typeof SUPPORTED_PALLETS)[number]
-export interface TPalletMap {
-  defaultPallet: TPallet
-  supportedPallets: TPallet[]
-}
-export type TPalletJsonMap = Record<TNode, TPalletMap>
-export type TEdJsonMap = Record<TNodeWithRelayChains, string | null>
 
-export interface TSerializedApiCall {
-  module: string
-  section: string
-  parameters: any[]
+export interface PolkadotXCMTransferInput {
+  api: ApiPromise
+  header: any
+  addressSelection: any
+  currencySelection: any
+  scenario: TScenario
+  currencySymbol: string | undefined
+  serializedApiCallEnabled?: boolean
 }
 
 export interface XTokensTransferInput {
@@ -77,6 +43,10 @@ export interface XTransferTransferInput {
   serializedApiCallEnabled?: boolean
 }
 
+export interface IPolkadotXCMTransfer {
+  transferPolkadotXCM: (input: PolkadotXCMTransferInput) => Extrinsic | TSerializedApiCall
+}
+
 export interface IXTokensTransfer {
   transferXTokens: (input: XTokensTransferInput) => Extrinsic | TSerializedApiCall
 }
@@ -85,14 +55,16 @@ export interface IXTransferTransfer {
   transferXTransfer: (input: XTransferTransferInput) => Extrinsic | TSerializedApiCall
 }
 
-export interface PolkadotXCMTransferInput {
-  api: ApiPromise
-  header: any
-  addressSelection: any
-  currencySelection: any
-  scenario: TScenario
-  currencySymbol: string | undefined
-  serializedApiCallEnabled?: boolean
+export type TScenario = 'ParaToRelay' | 'ParaToPara' | 'RelayToPara'
+
+export enum Version {
+  V1 = 'V1',
+  V3 = 'V3'
+}
+
+export enum Parents {
+  ONE = 1,
+  ZERO = 0
 }
 
 export type TAmount = string | number | bigint
@@ -148,47 +120,6 @@ export interface TRelayToParaCommonOptions extends TRelayToParaOptions {
   serializedApiCallEnabled?: boolean
 }
 
-export interface TOpenChannelOptions {
-  api: ApiPromise
-  origin: TNode
-  destination: TNode
-  maxSize: number
-  maxMessageSize: number
-}
-
-export interface TOpenChannelInternalOptions extends TOpenChannelOptions {
-  serializedApiCallEnabled?: boolean
-}
-
-export interface TCloseChannelOptions {
-  api: ApiPromise
-  origin: TNode
-  inbound: number
-  outbound: number
-}
-
-export interface TCloseChannelInternalOptions extends TCloseChannelOptions {
-  serializedApiCallEnabled?: boolean
-}
-
-export interface IPolkadotXCMTransfer {
-  transferPolkadotXCM: (input: PolkadotXCMTransferInput) => Extrinsic | TSerializedApiCall
-}
-
-export enum Version {
-  V1 = 'V1',
-  V3 = 'V3'
-}
-
-export enum Parents {
-  ONE = 1,
-  ZERO = 0
-}
-
-export type TMultiLocationHeader = {
-  [key in Version]?: TMultiLocation
-}
-
 export interface TCurrencySelection {
   id: {
     Concrete: TMultiLocation
@@ -204,6 +135,12 @@ export type TCurrencySelectionHeader = {
 
 export type TCurrencySelectionHeaderArr = {
   [key in Version]?: [TCurrencySelection]
+}
+
+export interface TSerializedApiCall {
+  module: string
+  section: string
+  parameters: any[]
 }
 
 export interface CheckKeepAliveOptions {
